requirements.txt
#esto sale usando pip freeze
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
flask-cors==3.0.10